+++ packages/api-client/src/http/HttpClient.ts
@@ -20,6 +20,7 @@
 import axios, {AxiosError, AxiosHeaders, AxiosInstance, AxiosRequestConfig, AxiosResponse} from 'axios';
 import axiosRetry, {isNetworkOrIdempotentRequestError, exponentialDelay} from 'axios-retry';
 import logdown from 'logdown';
+import {gzip} from 'pako';
 
 import {EventEmitter} from 'events';
 
@@ -135,9 +136,7 @@ export class HttpClient extends EventEmitter {
         config.headers = {
           ...config.headers,
           Authorization: `${token_type} ${access_token}`,
-          // There is a typedef error in axios 1.1.2. This has been fixed with 1.1.3 (that we cannot use as of now because of webpack build https://github.com/axios/axios/issues/5154)
-          // This `any` can be removed when we use axios > 1.1.3
-        } as any;
+        };
       }
     }
 
@@ -279,12 +278,21 @@ export class HttpClient extends EventEmitter {
   }
 
   public sendJSON<T>(config: AxiosRequestConfig, isSynchronousRequest: boolean = false): Promise<AxiosResponse<T>> {
+    const shouldGzipData =
+      process.env.NODE_ENV !== 'test' &&
+      !!config.data &&
+      ['post', 'put', 'patch'].includes(config.method?.toLowerCase() ?? '');
+
+    if (shouldGzipData) {
+      config.data = gzip(JSON.stringify(config.data));
+    }
+
     config.headers = {
       ...config.headers,
       'Content-Type': ContentType.APPLICATION_JSON,
-      // There is a typedef error in axios 1.1.2. This has been fixed with 1.1.3 (that we cannot use as of now because of webpack build https://github.com/axios/axios/issues/5154)
-      // This `any` can be removed when we use axios > 1.1.3
-    } as any;
+      'Content-Encoding': shouldGzipData ? 'gzip' : config.headers?.['Content-Encoding'],
+    };
+
     return this.sendRequest<T>(config, false, isSynchronousRequest);
   }
 
@@ -292,9 +300,7 @@ export class HttpClient extends EventEmitter {
     config.headers = {
       ...config.headers,
       'Content-Type': ContentType.APPLICATION_XML,
-      // There is a typedef error in axios 1.1.2. This has been fixed with 1.1.3 (that we cannot use as of now because of webpack build https://github.com/axios/axios/issues/5154)
-      // This `any` can be removed when we use axios > 1.1.3
-    } as any;
+    };
     return this.sendRequest<T>(config, false, false);
   }
 
@@ -305,9 +311,7 @@ export class HttpClient extends EventEmitter {
     config.headers = {
       ...config.headers,
       'Content-Type': ContentType.APPLICATION_PROTOBUF,
-      // There is a typedef error in axios 1.1.2. This has been fixed with 1.1.3 (that we cannot use as of now because of webpack build https://github.com/axios/axios/issues/5154)
-      // This `any` can be removed when we use axios > 1.1.3
-    } as any;
+    };
     return this.sendRequest<T>(config, false, isSynchronousRequest);
   }
 
@@ -318,9 +322,7 @@ export class HttpClient extends EventEmitter {
     config.headers = {
       ...config.headers,
       'Content-Type': ContentType.MESSAGES_MLS,
-      // There is a typedef error in axios 1.1.2. This has been fixed with 1.1.3 (that we cannot use as of now because of webpack build https://github.com/axios/axios/issues/5154)
-      // This `any` can be removed when we use axios > 1.1.3
-    } as any;
+    };
     return this.sendRequest<T>(config, false, isSynchronousRequest);
   }
 }