+++ packages/core/src/Account.ts
@@ -393,7 +393,6 @@ export class Account extends TypedEventEmitter<Events> {
       clientType === CryptoClientType.CORE_CRYPTO && this.isMlsEnabled()
         ? new MLSService(this.apiClient, cryptoClient.getNativeClient(), {
             ...this.cryptoProtocolConfig?.mls,
-            nbKeyPackages: this.nbPrekeys,
           })
         : undefined;
 
@@ -555,7 +554,7 @@ export class Account extends TypedEventEmitter<Events> {
 
     const handleMissedNotifications = async (notificationId: string) => {
       if (this.cryptoProtocolConfig?.mls && this.backendFeatures.supportsMLS) {
-        await this.service?.conversation.handleEpochMismatch();
+        await this.service?.conversation.handleConversationsEpochMismatch();
       }
       return onMissedNotifications(notificationId);
     };