+++ packages/core/src/conversation/ConversationService/ConversationService.test.ts
@@ -19,16 +19,29 @@
 
 import {ClientClassification, ClientType} from '@wireapp/api-client/lib/client';
 import {Conversation, ConversationProtocol, MLSConversation} from '@wireapp/api-client/lib/conversation';
+import {CONVERSATION_EVENT, ConversationMLSMessageAddEvent} from '@wireapp/api-client/lib/event';
+import {QualifiedId} from '@wireapp/api-client/lib/user';
 
 import {APIClient} from '@wireapp/api-client';
 import {GenericMessage} from '@wireapp/protocol-messaging';
 
 import {ConversationService, MessageSendingState} from '..';
 import {MLSService} from '../../messagingProtocols/mls';
+import {CoreCryptoMLSErrors} from '../../messagingProtocols/mls/MLSService/CoreCryptoMLSErrors';
 import {ProteusService} from '../../messagingProtocols/proteus';
 import * as MessagingProtocols from '../../messagingProtocols/proteus';
 import * as PayloadHelper from '../../test/PayloadHelper';
 import * as MessageBuilder from '../message/MessageBuilder';
+
+const createMLSMessageAddEventMock = (conversationId: QualifiedId): ConversationMLSMessageAddEvent => ({
+  data: '',
+  conversation: conversationId.id,
+  qualified_conversation: conversationId,
+  from: '',
+  senderClientId: '',
+  type: CONVERSATION_EVENT.MLS_MESSAGE_ADD,
+  time: '2023-08-21T06:47:43.387Z',
+});
 
 jest.mock('../../messagingProtocols/proteus', () => ({
   ...jest.requireActual('../../messagingProtocols/proteus'),
@@ -42,15 +55,6 @@ jest.mock('../message/messageSender', () => ({
   ...jest.requireActual('../message/messageSender'),
   sendMessage: jest.fn().mockImplementation(fn => fn()),
 }));
-
-const mockedMLSService = {
-  encryptMessage: () => {},
-  commitPendingProposals: () => Promise.resolve(),
-  getEpoch: () => Promise.resolve(),
-  joinByExternalCommit: jest.fn(),
-  registerConversation: jest.fn(),
-  wipeConversation: jest.fn(),
-} as unknown as MLSService;
 
 const mockedProteusService = {
   encryptGenericMessage: () => Promise.resolve(),
@@ -92,9 +96,22 @@ describe('ConversationService', () => {
       clientId: PayloadHelper.getUUID(),
     };
 
+    const mockedMLSService = {
+      encryptMessage: () => {},
+      commitPendingProposals: () => Promise.resolve(),
+      getEpoch: () => Promise.resolve(),
+      joinByExternalCommit: jest.fn(),
+      registerConversation: jest.fn(),
+      wipeConversation: jest.fn(),
+      handleMLSMessageAddEvent: jest.fn(),
+      conversationExists: jest.fn(),
+      isConversationEstablished: jest.fn(),
+    } as unknown as MLSService;
+
     const conversationService = new ConversationService(client, mockedProteusService, mockedMLSService);
 
     jest.spyOn(conversationService, 'joinByExternalCommit');
+    jest.spyOn(conversationService, 'emit');
 
     return [conversationService, {apiClient: client, mlsService: mockedMLSService}] as const;
   }
@@ -152,7 +169,7 @@ describe('ConversationService', () => {
     });
   });
 
-  describe('"handleEpochMismatch"', () => {
+  describe('handleConversationsEpochMismatch', () => {
     beforeEach(() => {
       jest.clearAllMocks();
     });
@@ -169,15 +186,17 @@ describe('ConversationService', () => {
     it('re-joins multiple not-established conversations', async () => {
       const [conversationService, {apiClient}] = buildConversationService();
 
-      const mlsConversation1 = createConversation(1, 'conversation1');
-      const mlsConversation2 = createConversation(1, 'conversation2');
+      const remoteEpoch = 1;
+
+      const mlsConversation1 = createConversation(remoteEpoch, 'conversation1');
+      const mlsConversation2 = createConversation(remoteEpoch, 'conversation2');
 
       const mockedDBResponse: Conversation[] = [mlsConversation1, mlsConversation2];
       jest.spyOn(apiClient.api.conversation, 'getConversationList').mockResolvedValueOnce({found: mockedDBResponse});
 
-      jest.spyOn(conversationService, 'isMLSConversationEstablished').mockResolvedValue(false);
+      jest.spyOn(conversationService, 'mlsGroupExistsLocally').mockResolvedValue(false);
 
-      await conversationService.handleEpochMismatch();
+      await conversationService.handleConversationsEpochMismatch();
       expect(conversationService.joinByExternalCommit).toHaveBeenCalledWith(mlsConversation1.qualified_id);
       expect(conversationService.joinByExternalCommit).toHaveBeenCalledWith(mlsConversation2.qualified_id);
     });
@@ -191,10 +210,10 @@ describe('ConversationService', () => {
       const mockedDBResponse: Conversation[] = [mlsConversation1, mlsConversation2];
       jest.spyOn(apiClient.api.conversation, 'getConversationList').mockResolvedValueOnce({found: mockedDBResponse});
 
-      jest.spyOn(conversationService, 'isMLSConversationEstablished').mockResolvedValue(true);
+      jest.spyOn(conversationService, 'mlsGroupExistsLocally').mockResolvedValue(true);
       jest.spyOn(mlsService, 'getEpoch').mockResolvedValue(2);
 
-      await conversationService.handleEpochMismatch();
+      await conversationService.handleConversationsEpochMismatch();
       expect(conversationService.joinByExternalCommit).toHaveBeenCalledWith(mlsConversation1.qualified_id);
       expect(conversationService.joinByExternalCommit).toHaveBeenCalledWith(mlsConversation2.qualified_id);
     });
@@ -207,32 +226,18 @@ describe('ConversationService', () => {
       const mockedDBResponse: Conversation[] = [mlsConversation];
       jest.spyOn(apiClient.api.conversation, 'getConversationList').mockResolvedValueOnce({found: mockedDBResponse});
 
-      jest.spyOn(conversationService, 'isMLSConversationEstablished').mockResolvedValueOnce(true);
+      jest.spyOn(conversationService, 'mlsGroupExistsLocally').mockResolvedValueOnce(true);
 
       jest.spyOn(mlsService, 'getEpoch').mockResolvedValueOnce(1);
+      jest.spyOn(mlsService, 'conversationExists').mockResolvedValueOnce(true);
 
-      await conversationService.handleEpochMismatch();
+      await conversationService.handleConversationsEpochMismatch();
       expect(conversationService.joinByExternalCommit).not.toHaveBeenCalled();
     });
   });
 
   describe('establishMLS1to1Conversation', () => {
-    it('successfully register an MLS group if it did not exist before', async () => {
-      const [conversationService, {mlsService}] = buildConversationService();
-
-      const mockGroupId = 'mock-group-id';
-      const selfUser = {user: {id: 'self-user-id', domain: 'staging.zinfra.io'}, client: 'self-user-client-id'};
-      const otherUserId = {id: 'other-user-id', domain: 'staging.zinfra.io'};
-
-      await conversationService.establishMLS1to1Conversation(mockGroupId, selfUser, otherUserId);
-
-      expect(mlsService.registerConversation).toHaveBeenCalledTimes(1);
-      expect(mlsService.registerConversation).toHaveBeenCalledWith(mockGroupId, [otherUserId, selfUser.user], selfUser);
-      expect(mlsService.joinByExternalCommit).not.toHaveBeenCalled();
-      expect(mlsService.wipeConversation).not.toHaveBeenCalled();
-    });
-
-    it('joins with external commit if epoch number is higher than 0', async () => {
+    it('only returns a conversation if a group is already established on backend and locally', async () => {
       const [conversationService, {apiClient, mlsService}] = buildConversationService();
 
       const mockConversationId = {id: 'mock-conversation-id', domain: 'staging.zinfra.io'};
@@ -241,23 +246,65 @@ describe('ConversationService', () => {
       const selfUser = {user: {id: 'self-user-id', domain: 'staging.zinfra.io'}, client: 'self-user-client-id'};
       const otherUserId = {id: 'other-user-id', domain: 'staging.zinfra.io'};
 
-      jest.spyOn(mlsService, 'registerConversation').mockRejectedValueOnce(undefined);
+      const remoteEpoch = 1;
+
       jest.spyOn(apiClient.api.conversation, 'getMLS1to1Conversation').mockResolvedValueOnce({
         qualified_id: mockConversationId,
         protocol: ConversationProtocol.MLS,
-        epoch: 1,
+        epoch: remoteEpoch,
         group_id: mockGroupId,
       } as unknown as MLSConversation);
+      jest.spyOn(mlsService, 'isConversationEstablished').mockResolvedValueOnce(true);
 
       await conversationService.establishMLS1to1Conversation(mockGroupId, selfUser, otherUserId);
 
-      expect(mlsService.registerConversation).toHaveBeenCalledTimes(1);
-      expect(mlsService.registerConversation).toHaveBeenCalledWith(mockGroupId, [otherUserId, selfUser.user], selfUser);
+      expect(mlsService.registerConversation).not.toHaveBeenCalled();
 
+      expect(conversationService.joinByExternalCommit).not.toHaveBeenCalled();
+    });
+
+    it('joins with an external commit if a group is already established on backend but not established locally', async () => {
+      const [conversationService, {apiClient, mlsService}] = buildConversationService();
+
+      const mockConversationId = {id: 'mock-conversation-id', domain: 'staging.zinfra.io'};
+      const mockGroupId = 'mock-group-id';
+
+      const selfUser = {user: {id: 'self-user-id', domain: 'staging.zinfra.io'}, client: 'self-user-client-id'};
+      const otherUserId = {id: 'other-user-id', domain: 'staging.zinfra.io'};
+
+      const remoteEpoch = 1;
+      const updatedEpoch = 2;
+
+      jest.spyOn(apiClient.api.conversation, 'getMLS1to1Conversation').mockResolvedValueOnce({
+        qualified_id: mockConversationId,
+        protocol: ConversationProtocol.MLS,
+        epoch: remoteEpoch,
+        group_id: mockGroupId,
+      } as unknown as MLSConversation);
+
+      // The 2nd request we make after joining the conversation with external commit
+      jest.spyOn(apiClient.api.conversation, 'getMLS1to1Conversation').mockResolvedValueOnce({
+        qualified_id: mockConversationId,
+        protocol: ConversationProtocol.MLS,
+        epoch: updatedEpoch,
+        group_id: mockGroupId,
+      } as unknown as MLSConversation);
+
+      jest.spyOn(mlsService, 'isConversationEstablished').mockResolvedValueOnce(false);
+      jest.spyOn(mlsService, 'joinByExternalCommit').mockResolvedValueOnce({events: [], time: ''});
+
+      const establishedConversation = await conversationService.establishMLS1to1Conversation(
+        mockGroupId,
+        selfUser,
+        otherUserId,
+      );
+
+      expect(mlsService.registerConversation).not.toHaveBeenCalled();
       expect(conversationService.joinByExternalCommit).toHaveBeenCalledWith(mockConversationId);
+      expect(establishedConversation.epoch).toEqual(updatedEpoch);
     });
 
-    it('retries to register mls group if epoch number is equal 0', async () => {
+    it('wipes the conversation and registers it if a group is not yet established on backend', async () => {
       const [conversationService, {apiClient, mlsService}] = buildConversationService();
 
       const mockConversationId = {id: 'mock-conversation-id', domain: 'staging.zinfra.io'};
@@ -266,19 +313,120 @@ describe('ConversationService', () => {
       const selfUser = {user: {id: 'self-user-id', domain: 'staging.zinfra.io'}, client: 'self-user-client-id'};
       const otherUserId = {id: 'other-user-id', domain: 'staging.zinfra.io'};
 
-      jest.spyOn(mlsService, 'registerConversation').mockRejectedValueOnce(undefined);
+      const remoteEpoch = 0;
+      const updatedEpoch = 1;
+
       jest.spyOn(apiClient.api.conversation, 'getMLS1to1Conversation').mockResolvedValueOnce({
         qualified_id: mockConversationId,
         protocol: ConversationProtocol.MLS,
-        epoch: 0,
+        epoch: remoteEpoch,
         group_id: mockGroupId,
       } as unknown as MLSConversation);
 
-      await conversationService.establishMLS1to1Conversation(mockGroupId, selfUser, otherUserId);
+      // The 2nd request we make after successfully registering a group
+      jest.spyOn(apiClient.api.conversation, 'getMLS1to1Conversation').mockResolvedValueOnce({
+        qualified_id: mockConversationId,
+        protocol: ConversationProtocol.MLS,
+        epoch: updatedEpoch,
+        group_id: mockGroupId,
+      } as unknown as MLSConversation);
 
+      jest.spyOn(mlsService, 'wipeConversation');
+
+      const establishedConversation = await conversationService.establishMLS1to1Conversation(
+        mockGroupId,
+        selfUser,
+        otherUserId,
+      );
+
+      expect(mlsService.wipeConversation).toHaveBeenCalledWith(mockGroupId);
+      expect(mlsService.registerConversation).toHaveBeenCalledTimes(1);
       expect(mlsService.registerConversation).toHaveBeenCalledWith(mockGroupId, [otherUserId, selfUser.user], selfUser);
+      expect(conversationService.joinByExternalCommit).not.toHaveBeenCalled();
+      expect(establishedConversation.epoch).toEqual(updatedEpoch);
+    });
+
+    it('retries to register the conversation after it has failed for the first time', async () => {
+      const [conversationService, {apiClient, mlsService}] = buildConversationService();
+
+      const mockConversationId = {id: 'mock-conversation-id', domain: 'staging.zinfra.io'};
+      const mockGroupId = 'mock-group-id';
+
+      const selfUser = {user: {id: 'self-user-id', domain: 'staging.zinfra.io'}, client: 'self-user-client-id'};
+      const otherUserId = {id: 'other-user-id', domain: 'staging.zinfra.io'};
+
+      const remoteEpoch = 0;
+      const updatedEpoch = 1;
+
+      jest.spyOn(apiClient.api.conversation, 'getMLS1to1Conversation').mockResolvedValueOnce({
+        qualified_id: mockConversationId,
+        protocol: ConversationProtocol.MLS,
+        epoch: remoteEpoch,
+        group_id: mockGroupId,
+      } as unknown as MLSConversation);
+
+      // The 2nd request we make when retrying to register the conversation
+      jest.spyOn(apiClient.api.conversation, 'getMLS1to1Conversation').mockResolvedValueOnce({
+        qualified_id: mockConversationId,
+        protocol: ConversationProtocol.MLS,
+        epoch: remoteEpoch,
+        group_id: mockGroupId,
+      } as unknown as MLSConversation);
+
+      // The 3rd request we make after successfully registering a group
+      jest.spyOn(apiClient.api.conversation, 'getMLS1to1Conversation').mockResolvedValueOnce({
+        qualified_id: mockConversationId,
+        protocol: ConversationProtocol.MLS,
+        epoch: updatedEpoch,
+        group_id: mockGroupId,
+      } as unknown as MLSConversation);
+
+      jest.spyOn(mlsService, 'registerConversation').mockRejectedValueOnce(undefined);
+      jest.spyOn(mlsService, 'wipeConversation');
+
+      const establishedConversation = await conversationService.establishMLS1to1Conversation(
+        mockGroupId,
+        selfUser,
+        otherUserId,
+      );
+
       expect(mlsService.wipeConversation).toHaveBeenCalledWith(mockGroupId);
       expect(mlsService.registerConversation).toHaveBeenCalledTimes(2);
+      expect(mlsService.registerConversation).toHaveBeenCalledWith(mockGroupId, [otherUserId, selfUser.user], selfUser);
+      expect(conversationService.joinByExternalCommit).not.toHaveBeenCalled();
+      expect(establishedConversation.epoch).toEqual(updatedEpoch);
+    });
+  });
+
+  describe('handleEvent', () => {
+    it('rejoins a MLS conversation if epoch mismatch detected when decrypting mls message', async () => {
+      const [conversationService, {apiClient, mlsService}] = buildConversationService();
+      const conversationId = {id: 'conversationId', domain: 'staging.zinfra.io'};
+      const mockGroupId = 'mock-group-id';
+
+      const mockMLSMessageAddEvent = createMLSMessageAddEventMock(conversationId);
+
+      jest
+        .spyOn(mlsService, 'handleMLSMessageAddEvent')
+        .mockRejectedValueOnce(new Error(CoreCryptoMLSErrors.WRONG_EPOCH));
+
+      const remoteEpoch = 5;
+      const localEpoch = 4;
+
+      jest.spyOn(mlsService, 'conversationExists').mockResolvedValueOnce(true);
+      jest.spyOn(mlsService, 'getEpoch').mockResolvedValueOnce(localEpoch);
+
+      jest.spyOn(apiClient.api.conversation, 'getConversation').mockResolvedValueOnce({
+        qualified_id: conversationId,
+        protocol: ConversationProtocol.MLS,
+        epoch: remoteEpoch,
+        group_id: mockGroupId,
+      } as unknown as Conversation);
+
+      await conversationService.handleEvent(mockMLSMessageAddEvent);
+
+      expect(conversationService.joinByExternalCommit).toHaveBeenCalledWith(conversationId);
+      expect(conversationService.emit).toHaveBeenCalledWith('MLSConversationRecovered', {conversationId});
     });
   });