+++ packages/core/src/messagingProtocols/mls/MLSService/CoreCryptoMLSErrors.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2022 Wire Swiss GmbH
+ * Copyright (C) 2023 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,5 +17,10 @@
  *
  */
 
-export * from './MLSService/';
-export * from './types';
+export enum CoreCryptoMLSErrors {
+  WRONG_EPOCH = 'Incoming message is for the wrong epoch',
+}
+
+export const isCoreCryptoMLSWrongEpochError = (error: unknown): boolean => {
+  return error instanceof Error && error.message === CoreCryptoMLSErrors.WRONG_EPOCH;
+};