+++ packages/core/src/messagingProtocols/mls/MLSService/MLSService.test.ts
@@ -70,6 +70,7 @@ describe('MLSService', () => {
       expect(apiClient.api.client.claimMLSKeyPackages).toHaveBeenCalledWith(
         selfUser.id,
         selfUser.domain,
+        '0x1',
         creator.client,
       );
       expect(mlsService.scheduleKeyMaterialRenewal).toHaveBeenCalledWith(groupId);
@@ -88,6 +89,46 @@ describe('MLSService', () => {
       jest.spyOn(mockCoreCrypto, 'conversationExists').mockResolvedValueOnce(true);
       await mlsService.wipeConversation(groupId);
       expect(mlsService.cancelKeyMaterialRenewal).toHaveBeenCalledWith(groupId);
+    });
+  });
+
+  describe('isConversationEstablished', () => {
+    it('returns false if conversation does not exist locally', async () => {
+      const mlsService = new MLSService(apiClient, mockCoreCrypto, {});
+
+      const groupId = 'mXOagqRIX/RFd7QyXJA8/Ed8X+hvQgLXIiwYHm3OQFc=';
+
+      jest.spyOn(mlsService, 'conversationExists').mockResolvedValueOnce(false);
+
+      const isEstablshed = await mlsService.isConversationEstablished(groupId);
+
+      expect(isEstablshed).toBe(false);
+    });
+
+    it('returns false if epoch number is 0', async () => {
+      const mlsService = new MLSService(apiClient, mockCoreCrypto, {});
+
+      const groupId = 'mXOagqRIX/RFd7QyXJA8/Ed8X+hvQgLXIiwYHm3OQFc=';
+
+      jest.spyOn(mlsService, 'conversationExists').mockResolvedValueOnce(true);
+      jest.spyOn(mlsService, 'getEpoch').mockResolvedValueOnce(0);
+
+      const isEstablshed = await mlsService.isConversationEstablished(groupId);
+
+      expect(isEstablshed).toBe(false);
+    });
+
+    it.each([1, 2, 100])('returns false if epoch number is 1 or more', async epoch => {
+      const mlsService = new MLSService(apiClient, mockCoreCrypto, {});
+
+      const groupId = 'mXOagqRIX/RFd7QyXJA8/Ed8X+hvQgLXIiwYHm3OQFc=';
+
+      jest.spyOn(mlsService, 'conversationExists').mockResolvedValueOnce(true);
+      jest.spyOn(mlsService, 'getEpoch').mockResolvedValueOnce(epoch);
+
+      const isEstablshed = await mlsService.isConversationEstablished(groupId);
+
+      expect(isEstablshed).toBe(true);
     });
   });
 });