+++ packages/core/src/util/numberToHex.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2022 Wire Swiss GmbH
+ * Copyright (C) 2023 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,5 +17,6 @@
  *
  */
 
-export * from './MLSService/';
-export * from './types';
+export const numberToHex = (number: number): string => {
+  return `0x${number.toString(16).toUpperCase()}`;
+};