+++ packages/bazinga64/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.1.5](https://github.com/wireapp/wire-web-packages/compare/bazinga64@6.1.4...bazinga64@6.1.5) (2023-09-06)
+
+**Note:** Version bump only for package bazinga64
+
+## [6.1.4](https://github.com/wireapp/wire-web-packages/compare/bazinga64@6.1.3...bazinga64@6.1.4) (2023-09-06)
+
+**Note:** Version bump only for package bazinga64
+
+## [6.1.3](https://github.com/wireapp/wire-web-packages/compare/bazinga64@6.1.2...bazinga64@6.1.3) (2023-09-04)
+
+**Note:** Version bump only for package bazinga64
+
 ## [6.1.2](https://github.com/wireapp/wire-web-packages/compare/bazinga64@6.1.1...bazinga64@6.1.2) (2023-08-10)
 
 **Note:** Version bump only for package bazinga64