+++ packages/promise-queue/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.3](https://github.com/wireapp/wire-web-packages/compare/@wireapp/promise-queue@2.2.2...@wireapp/promise-queue@2.2.3) (2023-09-06)
+
+**Note:** Version bump only for package @wireapp/promise-queue
+
+## [2.2.2](https://github.com/wireapp/wire-web-packages/compare/@wireapp/promise-queue@2.2.1...@wireapp/promise-queue@2.2.2) (2023-09-06)
+
+**Note:** Version bump only for package @wireapp/promise-queue
+
 ## [2.2.1](https://github.com/wireapp/wire-web-packages/compare/@wireapp/promise-queue@2.2.0...@wireapp/promise-queue@2.2.1) (2023-08-21)
 
 **Note:** Version bump only for package @wireapp/promise-queue